+++ src/controllers/PostConfigurationController.php
@@ -51,6 +51,28 @@ class PostConfigurationController extends AbstractJobController
         $connectorApiUrlRecord->value = $request->getBodyParam('connectorApiUrl');
         $connectorApiUrlRecord->save();
 
+        # enableEntriesForTargetSites
+        $enableEntriesForTargetSites = SettingRecord::findOne(['name' => 'enable_entries_for_target_sites']);
+        if (!$enableEntriesForTargetSites) {
+            $enableEntriesForTargetSites = new SettingRecord(['name' => 'enable_entries_for_target_sites']);
+        }
+        $enableEntriesForTargetSites->value = (int) $request->getBodyParam('enableEntriesForTargetSites');
+        $enableEntriesForTargetSites->save();
+
+        # copyEntriesSlugFromSourceToTarget
+        $copyEntriesSlugFromSourceToTarget = SettingRecord::findOne(
+            ['name' => 'copy_entries_slug_from_source_to_target']
+        );
+        if (!$copyEntriesSlugFromSourceToTarget) {
+            $copyEntriesSlugFromSourceToTarget = new SettingRecord(
+                ['name' => 'copy_entries_slug_from_source_to_target']
+            );
+        }
+        $copyEntriesSlugFromSourceToTarget->value = (int) $request->getBodyParam(
+            'copyEntriesSlugFromSourceToTarget'
+        );
+        $copyEntriesSlugFromSourceToTarget->save();
+
         $liltConfigDisabled = true;
         if (!empty($connectorApiKey) && !empty($connectorApiUrl)) {
             //is token valid