+++ src/utilities/Configuration.php
@@ -77,6 +77,15 @@ class Configuration extends Utility
         $connectorApiUrl = $connectorApiUrlRecord->value
             ?? \LiltConnectorSDK\Configuration::getDefaultConfiguration()->getHost();
 
+        $enableEntriesForTargetSitesRecord = SettingRecord::findOne(['name' => 'enable_entries_for_target_sites']);
+        $enableEntriesForTargetSites = (bool) ($enableEntriesForTargetSitesRecord->value
+            ?? false);
+
+        $copyEntriesSlugFromSourceToTargetRecord = SettingRecord::findOne(
+            ['name' => 'copy_entries_slug_from_source_to_target']
+        );
+        $copyEntriesSlugFromSourceToTarget = (bool) ($copyEntriesSlugFromSourceToTargetRecord->value ?? false);
+
         return Craft::$app->getView()->renderTemplate(
             'craft-lilt-plugin/_components/utilities/configuration.twig',
             [
@@ -88,6 +97,8 @@ class Configuration extends Utility
                 'connectorApiUrl' => $connectorApiUrl,
                 'formActionUrl' => UrlHelper::cpUrl('craft-lilt-plugin/settings/lilt-configuration'),
                 'liltConfigDisabled' => $liltConfigDisabled,
+                'enableEntriesForTargetSites' => $enableEntriesForTargetSites,
+                'copyEntriesSlugFromSourceToTarget' => $copyEntriesSlugFromSourceToTarget,
             ]
         );
     }